large overhaul of game service; add common domain
Commit that is over saturated. I was not sure what direction I wanted to take for this demo and made many changes in local throughout the past few days. The project is finally taking shape despite it all and needed to be committed.

Note: nearly everything is broken and unfinished; consider this commit to be nothing but a proof of concept scaffold.